final-pjt-back/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.4.5.2
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-auth==0.9.5
django-seed==0.2.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==4.1.3
idna==2.9
kiwisolver==1.3.1
lxml==4.6.1
matplotlib==3.3.3
nose==1.3.7
numpy==1.19.3
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.0.1
plotly==4.13.0
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.9
xmltodict==0.12.0
